Move DocFilesCheck.py checks to DocCheck.py
Move all DocFilesCheck.py checks to DocCheck as it logically belongs
there. Add comments, test cases and two test rpms:
  * install-file-in-docs-1.0-0.x86_64.rpm
  * doc-file-dependency-1.0.0-0.noarch.rpm

(#225 #156)
