+++ rpmlint/checks/DocCheck.py
@@ -1,19 +1,24 @@
 from pathlib import Path
 import stat
 
+import rpm
 from rpmlint.checks.AbstractCheck import AbstractCheck
+from rpmlint.helpers import byte_to_string
 
 
 class DocCheck(AbstractCheck):
     """
-    Checks for package documentation.
+    Package documentation checks.
     """
 
-    def check(self, pkg):
-        if pkg.isSource():
+    def check_binary(self, pkg):
+        if not pkg.docFiles():
             return
 
         self._check_executable_docs(pkg)
+        self._check_doc_file_dependencies(pkg)
+        self._check_unwanted_files(pkg)
+
         if not self._ignore_pkg(pkg.name):
             self._check_huge_docs(pkg)
 
@@ -41,6 +46,62 @@ class DocCheck(AbstractCheck):
             for name in ['README', 'NEWS', 'COPYING', 'AUTHORS', 'LICENCE', 'LICENSE']:
                 if path.name.lower() == name.lower():
                     self.output.add_info('E', pkg, 'executable-docs', f)
+
+    def _check_doc_file_dependencies(self, pkg):
+        """
+        Check if docfiles create additional dependencies in the package and
+        print a warning if so.
+        """
+        files = pkg.files()
+
+        reqs = {}
+        for fname, pkgfile in files.items():
+            reqs[fname] = [x[0] for x in pkgfile.requires]
+
+        core_reqs = {}  # dependencies of non-doc files
+        doc_reqs = {}   # dependencies of doc files
+
+        for dep in pkg.header.dsFromHeader():
+            # skip deps which were found by find-requires
+            if dep.Flags() & rpm.RPMSENSE_FIND_REQUIRES != 0:
+                continue
+            core_reqs[dep.N()] = []
+
+        # register things which are provided by the package
+        for i in pkg.header[rpm.RPMTAG_PROVIDES]:
+            core_reqs[byte_to_string(i)] = []
+        for i in files:
+            core_reqs[i] = []
+
+        for i in files:
+            if not reqs[i]:
+                continue  # skip empty dependencies
+            if i in pkg.docFiles():
+                target = doc_reqs
+            else:
+                target = core_reqs
+
+            for r in reqs[i]:
+                if r not in target:
+                    target[r] = []
+                target[r].append(i)
+
+        # go through the calculated requirements of the %doc files
+        for (dep, req_files) in doc_reqs.items():
+            if dep not in core_reqs:
+                for f in req_files:
+                    self.output.add_info('W', pkg, 'doc-file-dependency', f, dep)
+
+    def _check_unwanted_files(self, pkg):
+        """
+        Check if docfiles contain unwanted files.
+
+        Now it checks the presence of the INSTALL file that is often not
+        relevant for the built package.
+        """
+        for docfile in pkg.docFiles():
+            if docfile.endswith('/INSTALL'):
+                self.output.add_info('W', pkg, 'install-file-in-docs', docfile)
 
     def _check_huge_docs(self, pkg):
         """
