--- rpmlint/descriptions/DocFilesCheck.toml
@@ -1,11 +1,0 @@
-doc-file-dependency="""
-A file marked as %doc creates a possible additional dependency in the package.
-This is not wanted and may be caused by example scripts with executable bits
-set included in the package's documentation.
-"""
-
-install-file-in-docs="""
-A file whose name suggests that it contains installation instructions is
-included in the package. Such instructions are often not relevant for already
-installed packages.
-"""
